% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_ASEQ}
\alias{call_ASEQ}
\title{Format segmentation data for downstream analysis using CLONETv2}
\usage{
call_ASEQ(
  vcf = "",
  bin_path = "tools/ASEQ/binaries/linux64/ASEQ",
  bam = "",
  mrq = "",
  mbq = "",
  mdc = "",
  htperc = "",
  pht = "",
  mode = "",
  output_dir = "",
  threads = 3,
  verbose = FALSE
)
}
\arguments{
\item{vcf}{\link{REQUIRED} Path to vcf file.}

\item{bin_path}{\link{REQUIRED} Path to ASEQ binary. Default tools/ASEQ/binaries/linux64/ASEQ}

\item{bam}{\link{REQUIRED} Path to bam file.}

\item{mrq}{\link{OPTIONAL} Filter by read mapping quality >=. Default value 1. See ASEQ doc.}

\item{mbq}{\link{OPTIONAL} Filter by base quality >=. Default value 1. See ASEQ doc.}

\item{mdc}{\link{OPTIONAL} Filter by coverage per base >=. Default value 1. See ASEQ doc.}

\item{htperc}{\link{OPTIONAL} Heterozigosity test based on percentage. Default value 0.2. Only in GENOTYPE mode. See ASEQ doc.}

\item{pht}{\link{OPTIONAL} P-value for heterozigosity test. Default value 0.2. Only in GENOTYPE mode. See ASEQ doc.}

\item{mode}{\link{OPTIONAL} Usage mode. Default value PILEUP. See ASEQ doc.7}

\item{output_dir}{\link{OPTIONAL}  Path to the output directory.}

\item{threads}{\link{OPTIONAL} Number of threads to use. Default value 3.}

\item{verbose}{\link{OPTIONAL}  Enables progress messages. Default False.}

\item{seg_file}{\link{REQUIRED} Path/s to segmentation file/s. dir_segment and seg_file are mutually excluding.}

\item{dir_segment}{\link{REQUIRED} Path to directory with segmentation files.}

\item{pattern}{\link{FALSE} Pattern to use if directory for segmentation files is given.}

\item{cols_to_keep}{\link{DEFAULT==c(1,2,3,5)} Columns to keep from original segmentation bed files}

\item{output_name}{Name of the file to output.}
}
\description{
This function takes the segmentation data of multiple samples produced by any segment caller (mainly CNVkit) and generates
a single BED file with all the segment information (chr/start/end/log2) with an additional column for sample ID.
The default columns to select from original segmentation files are 1,2,3,5, which correspond to chr/start/end/log2 in CNVkit.
Use argument cols_to_keep to select other columns if needed or if order is different from different segment callers.
}
