% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_clonet}
\alias{call_clonet}
\title{This function is a wrapper around CLONET function}
\usage{
call_clonet(
  bin_path = "tools/bcftools/bcftools",
  bin_path2 = "tools/htslib/bgzip",
  bin_path3 = "tools/htslib/tabix",
  bin_path4 = "tools/ASEQ/binaries/linux64/ASEQ",
  bin_path5 = "tools/CLONET/CLONET.R",
  vcf = "",
  bam_dir = "",
  segment_data = "",
  patient_id = "",
  germ_pattern = "GL",
  sample_data = "",
  min_snp_cov = 10,
  min_nsnps = 10,
  min_seg_cov = 20,
  equal_betaThr = 0.9,
  max_homo_dels = 0.01,
  del_log_thr = c(-1, -0.25),
  alpha_par = 0.9,
  clonal_thr = 0.85,
  beta_thr = 0.85,
  stages = c(1, 2, 3, 4, 5, 6),
  comp_ref_map_bias = FALSE,
  beta_decimals = 3,
  ale_imb_thr = 0.5,
  beta_method = "STM",
  adm_method = "2D",
  verbose = FALSE,
  output_dir = "",
  jobs = 1,
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to bcftools binary. Default tools/bcftools/bcftools.}

\item{bin_path2}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path3}{Path to tabix binary. Default tools/htslib/tabix.}

\item{bin_path4}{Path to ASEQ binary. Default tools/ASEQ/binaries/linux64/ASEQ}

\item{bin_path5}{Path to CLONET. Default tools/CLONET/CLONET.R}

\item{vcf}{\link{REQUIRED} Path to VCF file with heterozygous SNPs.}

\item{bam_dir}{\link{REQUIRED} Path to directory with BAM files.}

\item{segment_data}{\link{REQUIRED} Path to file with formated segment data.}

\item{patient_id}{Patient ID. Default Patient}

\item{germ_pattern}{Germline pattern. Default GL}

\item{sample_data}{Path to files with sample info}

\item{min_snp_cov}{Minimum tumor coverage for informative SNPs. Default 10.}

\item{min_nsnps}{Minimum number of SNPs per segment. Default 10.}

\item{min_seg_cov}{Minimum segment coverage. Default 20.}

\item{equal_betaThr}{Minimum value of beta above which the two alleles are present in the same number. Default 0.9}

\item{max_homo_dels}{Homozygous deletions threshold. Default 0.01}

\item{del_log_thr}{Parameters of a valid deletion used to compute Adm.global. Default c(-1,-0.25)}

\item{alpha_par}{Percentage of used deletions to compute Adm.global varibility interval. Default 0.9}

\item{clonal_thr}{Clonality value threshold. Default 0.85}

\item{beta_thr}{Beta value threshold. Default 0.85}

\item{stages}{Analysis stages to run trough.Default c(1,2,3,4,5,6)}

\item{comp_ref_map_bias}{Compute reference mapping bias and to adjust beta estimation. Default FALSE}

\item{beta_decimals}{Number of beta value decimals to report. Default 3}

\item{beta_method}{Method for beta estimation. Default STM. Options STM/GB}

\item{adm_method}{Method for admixture estimation. Default 2D. Options 1D/2D}

\item{verbose}{Enables progress messages. Default FALSE.}

\item{output_dir}{Path to the output directory.}

\item{jobs}{Number of Samples to analyze in parallel. Default 1}

\item{threads}{Number of threads per job to use. Default 3}
}
\description{
This function takes multiple parameters and perform the following analysis:
}
