% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_fings}
\alias{call_fings}
\title{Filters SNVs from any variant caller to remove false positives (FP) using FiNGS}
\usage{
call_fings(
  bin_path = "tools/fings/FiNGS.py",
  tumor_bam = "",
  normal_bam = "",
  tumor_vcf = "",
  ref_genome = "",
  output_dir = "",
  max_depth = 1000,
  pass_in = TRUE,
  pass_out = FALSE,
  verbose = FALSE,
  param = "tools/fings/icgc_filter_parameters.txt"
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk binary. Default tools/gatk/gatk.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{max_depth}{\link{OPTIONAL} Maximum number of reads.Reads beyond this depth will be ignored}

\item{pass_in}{\link{OPTIONAL} Only input variants with a PASS. Default TRUE.}

\item{pass_out}{\link{OPTIONAL} Only output variants with a PASS. Default FALSE.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{param}{\link{REQUIRED} Path to file with filter params.}

\item{bam_dir}{\link{REQUIRED} Path to directory with BAM files.}

\item{vcf_dir}{\link{REQUIRED} Path to directory with VCF files.}

\item{patient_id}{\link{REQUIRED} Patient ID to analyze. Has to be in file names to subselect samples.}

\item{germ_pattern}{\link{REQUIRED} Pattern used to identify germline samples. Ex GL}

\item{threads}{\link{OPTIONAL} Number of threads. Default 3}
}
\description{
This function takes a paths of a pair of matched tumor/normal BAMs and the tumors' VCF.
Then it calls FiNGS to calculate metrics based on the BAM files, as well as providing filtering
for FP variants called by the variant caller.
}
