% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_fings_parallel}
\alias{call_fings_parallel}
\title{Filters SNVs from any variant caller to remove false positives (FP) using FiNGS on parallel}
\usage{
call_fings_parallel(
  bin_path = "tools/fings/FiNGS.py",
  bam_dir = "",
  vcf_dir = "",
  patient_id = "",
  germ_pattern = "GL",
  ref_genome = "",
  max_depth = 1000,
  param = "tools/fings/icgc_filter_parameters.txt",
  threads = 3,
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk binary. Default tools/gatk/gatk.}

\item{bam_dir}{\link{REQUIRED} Path to directory with BAM files.}

\item{vcf_dir}{\link{REQUIRED} Path to directory with VCF files.}

\item{patient_id}{\link{REQUIRED} Patient ID to analyze. Has to be in file names to subselect samples.}

\item{germ_pattern}{\link{REQUIRED} Pattern used to identify germline samples. Ex GL}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{max_depth}{\link{OPTIONAL} Maximum number of reads.Reads beyond this depth will be ignored}

\item{param}{\link{REQUIRED} Path to file with filter params.}

\item{threads}{\link{OPTIONAL} Number of threads. Default 3}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
This function takes a path to a directory with BAM files and a patient ID. Then it subsets all BAM
files specific to the patient, identifying between cancer and normal samples using the germline identifier.
It also takes a directory with vcf files and subsets them based on the sample ID.
Afterwards, it calls FiNGS to calculates metrics based on BAM files, providing filtering of FP.
}
