% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_HaplotypeCaller}
\alias{call_HaplotypeCaller}
\title{Wrapper for GATK HaplotypeCaller for Germline Variant calling}
\usage{
call_HaplotypeCaller(
  bin_path = "tools/gatk/gatk",
  bin_path2 = "tools/bcftools/bcftools",
  bin_path3 = "tools/htslib/bgzip",
  bin_path4 = "tools/htslib/tabix",
  normal_bam = "",
  ref_genome = "",
  region = "",
  output_dir = "",
  resources = "",
  patient_id = "",
  info_key = "CNN_1D",
  snp_tranche = 99.95,
  indel_tranche = 99.4,
  keep_previous_filters = FALSE,
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to GATK binary. Default tools/gatk/gatk}

\item{bin_path2}{Path to bcftools binary. Default tools/bcftools/bcftools.}

\item{bin_path3}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path4}{Path to tabix binary. Default tools/htslib/tabix.}

\item{normal_bam}{\link{REQUIRED} Path to BAM file}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{region}{\link{OPTIONAL} Genomic region/regions to analyze. In samtools format.}

\item{output_dir}{\link{OPTIONAL} Path to output dir}

\item{resources}{\link{OPTIONAL} Path to resources for variant filtering}

\item{patient_id}{\link{OPTIONAL} Patient ID after which to name the output. If not given bam file name will be used.}

\item{info_key}{\link{OPTIONAL} Annotation column to select. Default CNN_D1}

\item{snp_tranche}{\link{OPTIONAL} SNP tranche filter value. Default 99.95}

\item{indel_tranche}{\link{OPTIONAL} Indel tranche filter value. Default 99.4}

\item{keep_previous_filters}{\link{OPTIONAL} Keep previous filters in VCF. Default False}

\item{verbose}{\link{Optional} Enables progress messages. Default False}

\item{threads}{\link{OPTIONAL} Number of threads per job. Default 3}

\item{output_name}{\link{OPTIONAL} Name of the sample to output}
}
\description{
This function takes an annotated VCF with CNN_D1 or CNN_D2 scores and filter the
variants based on set threshold.Cuttently only single sample implementation
}
