% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_mutect2_parallel}
\alias{call_mutect2_parallel}
\title{Variant calling using MuTECT2 on parallel per genomic region}
\usage{
call_mutect2_parallel(
  bin_path = "tools/gatk/gatk",
  bin_path2 = "tools/bcftools/bcftools",
  bin_path3 = "tools/htslib/bgzip",
  bin_path4 = "tools/htslib/tabix",
  tumor_bam = "",
  normal_bam = "",
  ref_genome = "",
  germ_resource = "",
  pon = "",
  output_dir = "",
  output_name = "",
  region_bed = "",
  threads = 3,
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk binary. Default path tools/gatk/gatk.}

\item{bin_path2}{\link{REQUIRED} Path to bcftools binary. Default path tools/bcftools/bcftools.}

\item{bin_path3}{\link{REQUIRED} Path to tabix binary. Default path tools/htslib/tabix.}

\item{bin_path4}{\link{REQUIRED} Path to bgzip binary. Default path tools/htslib/bgzip.}

\item{tumor_bam}{\link{REQUIRED} Path to tumor bam file/s.}

\item{normal_bam}{\link{OPTIONAL} Path to germline bam file/s.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{germ_resource}{\link{REQUIRED} Path to germline resources vcf file.}

\item{pon}{\link{OPTIONAL} Path to panel of normal.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of one of the samples will be used.}

\item{region_bed}{\link{REQUIRED} Path to bed file with regions to analyze.}

\item{threads}{\link{OPTIONAL} Number of threads. Default 3}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
This function calls somatic variants in a pair of tumor-normal matched samples, or
just in a tumor sample if no matched sample is not available.
}
