% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_platypus}
\alias{call_platypus}
\title{Variant calling using Platypus}
\usage{
call_platypus(
  bin_path = "tools/platypus/Platypus.py",
  bin_path2 = "tools/htslib/bgzip",
  bin_path3 = "tools/htslib/tabix",
  tumor_bam = "",
  normal_bam = "",
  ref_genome = "",
  vcf_overlay = "",
  output_dir = "",
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to fastQC executable. Default path tools/platypus/Platypus.py.}

\item{bin_path2}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path3}{Path to tabix binary. Default tools/htslib/tabix.}

\item{tumor_bam}{Path to tumor bam file.}

\item{normal_bam}{Path to germline bam file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{vcf_overlay}{Path to vcf overlay to use as source.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads to use. Default 3.}
}
\description{
This function calls somatic variants in a pair of tumor-normal matched samples, or
just in a tumor sample if no matched sample is not available.
}
