% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_segments}
\alias{call_segments}
\title{Call segments for panel, exome and WGS data using CNVkit
This function takes a single or multiple tumor samples and a single or multiple normal samples,
creates a reference of a pool of normal samples, estimates the coverage and normalizes its log2 ratio using gc and accessibility information.
Ultimately, it calls the segments of binned data and identifies the state of CNA.}
\usage{
call_segments(
  bin_path = "~/tools/cnvkit/cnvkit.py",
  tumor_samples = "",
  normal_samples = "",
  pool_ref = "",
  targets = "",
  fasta = "",
  access = "",
  ref_output = "",
  output_dir = "",
  diagram = TRUE,
  scatter = TRUE,
  threads = 3,
  verbose = FALSE,
  male = TRUE
)
}
\arguments{
\item{bin_path}{Path to cnvkit binary.}

\item{tumor_samples}{\link{REQUIRED} Path to tumor samples bam files. For multiple samples pass as vector with their paths.}

\item{normal_samples}{\link{OPTIONAL} Path to normal samples bam files. For multiple samples pass as vector with their paths. Used to create a reference.}

\item{pool_ref}{\link{OPTIONAL} Path to reference of normals. Needs to be generated beforehand from normals/normals}

\item{targets}{\link{OPTIONAL} Path to exome capture targets. Required if targeted or panel.}

\item{access}{\link{OPTIONAL} Path to bed with accessibility information.}

\item{ref_output}{\link{OPTIONAL} Name of the reference file output.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{diagram}{\link{DEFAULT==TRUE} Plot diagram of segments.}

\item{scatter}{\link{DEFAULT==TRUE} Plot scatter plot of segments.}

\item{threads}{\link{DEFAULT==3} Number of threads to use.}

\item{verbose}{\link{DEFAULT==FALSE} Enables progress messages.}

\item{male}{\link{DEFAULT==TRUE} Reference is a male.}
}
\description{
Call segments for panel, exome and WGS data using CNVkit
This function takes a single or multiple tumor samples and a single or multiple normal samples,
creates a reference of a pool of normal samples, estimates the coverage and normalizes its log2 ratio using gc and accessibility information.
Ultimately, it calls the segments of binned data and identifies the state of CNA.
}
