% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_sv}
\alias{call_sv}
\title{Structural variant calling using svaba}
\usage{
call_sv(
  bin_path = "tools/svaba/bin/svaba",
  tumor_bam = "",
  normal_bam = "",
  ref_genome = "",
  threads = 3,
  output_name = "",
  targets = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to svaba binary executable. Default path tools/svaba/svaba.}

\item{tumor_bam}{Path to tumor bam file.}

\item{normal_bam}{Path to germline bam file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{threads}{Number of threads to use. Default 3.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first sample in alphanumerical order will be used.}

\item{targets}{BED file with capture target regions.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function calls structural variants in a pair of tumor-normal matched samples,
using svaba
}
