% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_sv_parallel}
\alias{call_sv_parallel}
\title{Structural variant calling using svaba in parallel}
\usage{
call_sv_parallel(
  bin_path = "tools/svaba/bin/svaba",
  targets = "",
  bam_dir = "",
  patient_id = "",
  germ_pattern = "GL",
  ref_genome = "",
  jobs = 3,
  threads = 1,
  par_type = "Joint",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to svaba binary executable. Default path tools/svaba/svaba.}

\item{targets}{\link{OPTIONAL} BED file with capture target regions.}

\item{bam_dir}{\link{REQUIRED}  Path to bam files.}

\item{patient_id}{\link{REQUIRED} Patient ID patter to filter samples with.}

\item{germ_pattern}{\link{OPTIONAL} Patient ID pattern to filter samples with.}

\item{ref_genome}{\link{REQUIRED} Germline pattern to filter samples with}

\item{jobs}{\link{OPTIONAL} Number of paired jobs to run. Only when par_type is Paired. Default 3.}

\item{threads}{\link{OPTIONAL} Number of threads to use per job. Default 1.}

\item{par_type}{\link{OPTIONAL} Parallelization type. Joint: Runs a single job with multiple samples jointly | Paired: Runs a pair of tumor-normal through multiple jobs.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}
}
\description{
This function calls structural variants in a pair of tumor-normal matched samples pairs
using svaba in parallel
}
