% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_variants_strelka}
\alias{call_variants_strelka}
\title{Call Somatic/Germline single nucleotide variants using Strelka
This function takes a pair of matched samples or single | multiple germline variants
and calls variants on them. To specify the variant calling mode supply the right strelka binary.
It is recommended to supply a vcf with indel candidates. This can be generated using MANTA workflow}
\usage{
call_variants_strelka(
  bin_path = "tools/strelka-2.9.10/build/bin/configureStrelkaSomaticWorkflow.py",
  bin_path2 = "tools/manta-1.6.0/build/bin/configManta.py",
  bin_path3 = "tools/bcftools/bcftools",
  bin_path4 = "tools/htslib/bgzip",
  bin_path5 = "tools/htslib/tabix",
  tumor_bam = "",
  normal_bam = "",
  ref_genome = "",
  output_dir = "",
  patient_id = "",
  verbose = FALSE,
  targeted = FALSE,
  threads = 3,
  exec_options = "local"
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to strelka binary. Somatic or Germline.}

\item{bin_path2}{\link{REQUIRED} Path to manta binary. Somatic or Germline.}

\item{bin_path3}{\link{REQUIRED} Path to bcftools binary.}

\item{bin_path4}{\link{REQUIRED} Path to bgzip binary.}

\item{bin_path5}{\link{REQUIRED} Path to tabix binary.}

\item{tumor_bam}{\link{REQUIRED} Path to tumor  bam file.}

\item{normal_bam}{\link{OPTIONAL} Path to normal samples bam files.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{patient_id}{\link{OPTIONAL} Patient ID to name files after. If not given tumor_bam file name will be used, and if this is not given then normal_bam file will be used.}

\item{verbose}{\link{DEFAULT==FALSE} Enables progress messages.}

\item{targeted}{\link{OPTIONAL} If exome/capture method. Default FALSE}

\item{threads}{\link{OPTIONAL} Number of threads per job. Default 3}

\item{exec_options}{\link{OPTIONAL} Type of execution. local (Single node) / sge (multiple nodes). Default local.}
}
\description{
Call Somatic/Germline single nucleotide variants using Strelka
This function takes a pair of matched samples or single | multiple germline variants
and calls variants on them. To specify the variant calling mode supply the right strelka binary.
It is recommended to supply a vcf with indel candidates. This can be generated using MANTA workflow
}
