% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_variants_strelka_parallel}
\alias{call_variants_strelka_parallel}
\title{Call Somatic/Germline variants using Strelka in parallel
This function takes a pair of matched samples or single | multiple germline variants
and calls variants on them. To specify the variant calling mode supply the right strelka binary.
It is recommended to supply a vcf with indel candidates. This can be generated using MANTA workflow.
This function parallelizes the number of jobs that can be run.}
\usage{
call_variants_strelka_parallel(
  bin_path = "~/tools/strelka-2.9.10/build/bin/configureStrelkaSomaticWorkflow.py",
  bam_dir = "",
  ref_genome = "",
  indel_candidates = "",
  output_dir = "",
  verbose = FALSE,
  targeted = FALSE,
  jobs = 1,
  threads = 3,
  exec_options = "local"
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to strelka binary. Somatic or Germline.}

\item{bam_dir}{\link{REQUIRED} Path to directory with BAM files.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome.}

\item{indel_candidates}{\link{OPTIONAL} Path to indel candidates produced by MANTA.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{DEFAULT==FALSE} Enables progress messages.}

\item{targeted}{\link{OPTIONAL} If exome/capture method. Default FALSE}

\item{jobs}{\link{OPTIONAL} Number of jobs. Default 1}

\item{threads}{\link{OPTIONAL} Number of threads per job. Default 3}

\item{exec_options}{\link{OPTIONAL} Type of execution. local (Single node) / sge (multiple nodes). Default local.}
}
\description{
Call Somatic/Germline variants using Strelka in parallel
This function takes a pair of matched samples or single | multiple germline variants
and calls variants on them. To specify the variant calling mode supply the right strelka binary.
It is recommended to supply a vcf with indel candidates. This can be generated using MANTA workflow.
This function parallelizes the number of jobs that can be run.
}
