% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_and_unzip}
\alias{check_and_unzip}
\title{Check and unzip gz compressed file}
\usage{
check_and_unzip(file = "", preserve = TRUE, verbose = FALSE)
}
\arguments{
\item{file}{\link{REQUIRED} Path to file}

\item{preserve}{\link{OPTIONAL} Preserve original file. Default TRUE}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
This function takes a file checks if its compressed and uncompresses if true.
}
