% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{CNNScoreVariants}
\alias{CNNScoreVariants}
\title{Annotate a VCF with scores from a Convolutional Neural Network (CNN) {Mutect2}}
\usage{
CNNScoreVariants(
  bin_path = "tools/gatk/gatk",
  vcf = "",
  ref_genome = "",
  bam = "",
  output_dir = "",
  output_name = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to GATK binary. Default tools/gatk/gatk}

\item{vcf}{\link{REQUIRED} Path to VCF file to annotate.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome.}

\item{bam}{\link{OPTIONAL} Path to BAM file.}

\item{output_dir}{\link{OPTIONAL} Path to BAM file.}

\item{output_name}{\link{OPTIONAL} Default output name. If not supplied it will use the vcf file name.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
This function takes a vcf and then annotates it with CNN scores.
If a BAM files is supplied a 2D model will be applied, otherwise a 1D model will be applied.
}
