% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compress_and_index_vcf}
\alias{compress_and_index_vcf}
\title{Compress and index vcf file}
\usage{
compress_and_index_vcf(
  bin_path = "tools/htslib/bgzip",
  bin_path2 = "tools/htslib/tabix",
  vcf = "",
  output_dir = ""
)
}
\arguments{
\item{vcf}{\link{REQUIRED} Path to VCF to compress}

\item{output_dir}{\link{OPTIONAL} Path to output directory}
}
\description{
This function takes an uncompressed vcf file and uses bgzip and tabix to
compress and index the file while preserving the original file
}
