% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{estimate_contamination}
\alias{estimate_contamination}
\title{VCF contamination estimation using GATK}
\usage{
estimate_contamination(
  pileup_table_tumor = "",
  bin_path = "tools/gatk/gatk",
  pileup_table_normal = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{pileup_table_tumor}{\link{REQUIRED} Path to pileup_table for tumor sample generated using pileup_summary}

\item{bin_path}{\link{REQUIRED} Path to gatk binary. Default tools/gatk/gatk.}

\item{pileup_table_normal}{\link{OPTIONAL} Path to pileup_table for matched normal sample generated using pileup_summary}

\item{verbose}{\link{OPTIONAL}  Enables progress messages. Default False.}

\item{output_dir}{\link{OPTIONAL}  Path to the output directory.}
}
\description{
This function estimates sample contamination for further variant filtering
}
