% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{estimate_contamination_parallel}
\alias{estimate_contamination_parallel}
\title{VCF contamination estimation using GATK}
\usage{
estimate_contamination_parallel(
  bin_path = "tools/gatk/gatk",
  bam_dir = "",
  germ_pattern = "GL",
  patient_id = "",
  db = "",
  interval = "",
  verbose = FALSE,
  output_dir = "",
  threads = 3
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk binary. Default tools/gatk/gatk.}

\item{bam_dir}{\link{REQUIRED} Path to directory with BAM files}

\item{germ_pattern}{\link{REQUIRED} Pattern to match normal samples.}

\item{patient_id}{\link{REQUIRED} Pattern to match patient specific samples.}

\item{db}{\link{REQUIRED} Path to vcf with known variants.}

\item{interval}{\link{REQUIRED} Path to vcf with intervals to analyze.}

\item{verbose}{\link{OPTIONAL}  Enables progress messages. Default False.}

\item{output_dir}{\link{OPTIONAL}  Path to the output directory.}

\item{threads}{\link{OPTIONAL}  Number of threads to use. Default 3.}
}
\description{
This function estimates sample contamination for further variant filtering
}
