% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fix_mnps}
\alias{fix_mnps}
\title{VCF fix MNPs}
\usage{
fix_mnps(
  vcf = "",
  bin_path = "tools/bcftools/bcftools",
  bin_path2 = "tools/htslib/bgzip",
  bin_path3 = "tools/htslib/tabix",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{vcf}{Path to unfiltered vcf file.}

\item{bin_path}{Path to bcftools binary. Default tools/bcftools/bcftools.}

\item{bin_path2}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path3}{Path to tabix binary. Default tools/htslib/tabix.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
Collapse multiple lines with the same genomic position into a single line in vcf
}
