% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{format_PM_analysis}
\alias{format_PM_analysis}
\title{Format vcf for downstream point mutatation analysis.}
\usage{
format_PM_analysis(
  bin_path = "tools/bcftools/bcftools",
  vcf = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to ASEQ binary. Default tools/bcftools/bcftools}

\item{vcf}{\link{REQUIRED} Path to vcf file.}

\item{output_dir}{\link{OPTIONAL}  Path to the output directory.}

\item{verbose}{\link{OPTIONAL}  Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to use. Default value 3.}
}
\description{
This function takes a VCF file and outputs a tab delimited file with data for downstream analysis in CLONETv2 point mutations analysis.
}
