% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{format_PM_data}
\alias{format_PM_data}
\title{Format vcf for downstream point mutatation analysis.}
\usage{
format_PM_data(
  bin_path = "tools/bcftools/bcftools",
  vcf_dir = "",
  pattern = "",
  vcf = "",
  output_dir = "",
  output_name = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to ASEQ binary. Default tools/bcftools/bcftools}

\item{vcf_dir}{\link{REQUIRED} Path to vcf file directory.}

\item{pattern}{\link{OPTIONAL} Pattern to filter vcfs by.}

\item{vcf}{\link{OPTIONAL} Path to vcf file. Only if vcf directory not provided.}

\item{output_dir}{\link{OPTIONAL}  Path to the output directory.}

\item{output_name}{\link{OPTIONAL}  Name of output file.}

\item{verbose}{\link{OPTIONAL}  Enables progress messages. Default False.}
}
\description{
This function takes a VCF file and outputs a tab delimited file with data for downstream analysis in CLONETv2 point mutations analysis.
}
