% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{format_segment_data}
\alias{format_segment_data}
\title{Format segmentation data for downstream analysis using CLONETv2}
\usage{
format_segment_data(
  seg_file = "",
  dir_segment = "",
  pattern = "",
  cols_to_keep = "",
  output_dir = "",
  output_name = "",
  verbose = FALSE
)
}
\arguments{
\item{seg_file}{\link{REQUIRED} Path/s to segmentation file/s. dir_segment and seg_file are mutually excluding.}

\item{dir_segment}{\link{REQUIRED} Path to directory with segmentation files.}

\item{pattern}{\link{FALSE} Pattern to use if directory for segmentation files is given.}

\item{cols_to_keep}{\link{DEFAULT==c(1,2,3,5)} Columns to keep from original segmentation bed files}

\item{output_dir}{Path to the output directory.}

\item{output_name}{Name of the file to output.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes the segmentation data of multiple samples produced by any segment caller (mainly CNVkit) and generates
a single BED file with all the segment information (chr/start/end/log2) with an additional column for sample ID.
The default columns to select from original segmentation files are 1,2,3,5, which correspond to chr/start/end/log2 in CNVkit.
Use argument cols_to_keep to select other columns if needed or if order is different from different segment callers.
}
