% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_pileup_summary}
\alias{get_pileup_summary}
\title{BAM pileup summary for known sites using GATK}
\usage{
get_pileup_summary(
  bam = "",
  bin_path = "tools/gatk/gatk",
  db = "",
  interval = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bam}{\link{REQUIRED} Path to BAM file.}

\item{bin_path}{\link{REQUIRED} Path to gatk binary. Default tools/gatk/gatk.}

\item{db}{\link{REQUIRED} Path to vcf with known variants.}

\item{interval}{\link{REQUIRED} Path to vcf with intervals to analyze.}

\item{verbose}{\link{OPTIONAL}  Enables progress messages. Default False.}

\item{output_dir}{\link{OPTIONAL}  Path to the output directory.}
}
\description{
This function generates a pileup summary that is further used in estimating contamination
}
