% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{learn_orientation}
\alias{learn_orientation}
\title{Learn Read Orientation Model (Mutect2)}
\usage{
learn_orientation(
  bin_path = "tools/gatk/gatk",
  f1r2 = "",
  f1r2_dir = "",
  output_name = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to GATK binary. Default tools/htslib/tabix.}

\item{f1r2}{\link{OPTIONAL} Path to f1r2 file. Only if f1r2_dir is not given.}

\item{f1r2_dir}{\link{OPTIONAL} Path to f1r2 file dir. Only if f1r2 is not given.}

\item{verbose}{\link{Optional} Enables progress messages. Default False.}
}
\description{
This function takes a f1r2 read information and generates an orientation model
to be used for further filtering steps.
}
