% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{merge_maf}
\alias{merge_maf}
\title{This function merges a group of files in MAF format
This function takes the path to the directory with MAF files and
concatenates them into a single file}
\usage{
merge_maf(
  maf_dir = "",
  recursive = FALSE,
  pattern = ".maf",
  threads = 3,
  generate_file = TRUE,
  output_name = "merged",
  output_dir = ""
)
}
\arguments{
\item{maf_dir}{Path to maf file dir}

\item{recursive}{Reads all files in the directory recursively. Default TRUE}

\item{pattern}{Pattern to identify MAF files with. Default "maf"}

\item{threads}{Number of threads to use.}

\item{generate_file}{Generate an output file. Default TRUE}

\item{output_name}{Output file name without extension. Default "merged"}

\item{output_dir}{Path to output directory.}
}
\description{
This function merges a group of files in MAF format
This function takes the path to the directory with MAF files and
concatenates them into a single file
}
