% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot_allelic_imbalance}
\alias{plot_allelic_imbalance}
\title{This function generates a plot of allelic imbalance from CLONET output}
\usage{
plot_allelic_imbalance(
  clonet_dir = "",
  sample_data = "",
  output_dir = "",
  gene_data = ""
)
}
\arguments{
\item{clonet_dir}{Path to clonet output directory}

\item{sample_data}{Path to file with sample info}

\item{output_dir}{Path to output directory.}

\item{gene_data}{Path to file with gene info}
}
\description{
This function takes the path to the directory with CLONET output, as well as a tab separated file with
sample and gene info and generates a plot of celularity and ploidy levels in the samples
}
