% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot_celullarity}
\alias{plot_celullarity}
\title{This function generates a plot of ploidy and celularity levels from CLONET data}
\usage{
plot_celullarity(clonet_dir = "", sample_data = "", output_dir = "")
}
\arguments{
\item{clonet_dir}{Path to clonet output directory}

\item{sample_data}{Path to file with sample info}

\item{output_dir}{Path to output directory.}
}
\description{
This function takes the path to the directory with CLONET output, as well as a tab separated file with
sample info and generates a plot of celularity and ploidy levels in the samples
}
