% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot_evolutionary_distance}
\alias{plot_evolutionary_distance}
\title{This function generates a plot of reltionship between samples based on Hammering Distance
computed as the number of bins thats differ between samples
This function takes the path to the directory with CLONET output, as well as a tab separated file with
sample info and generates a plot of celularity and ploidy levels in the samples}
\usage{
plot_evolutionary_distance(
  cn_call_data = "",
  sample_data = "",
  ref_bins = "",
  output_dir = "",
  threads = 3
)
}
\arguments{
\item{cn_call_data}{Path to segment data with cn calls}

\item{sample_data}{Path to file with sample info}

\item{ref_bins}{Path to reference genome binned into 100kb/500kb bin}

\item{output_dir}{Path to output directory.}

\item{threads}{Number of threads to use.}
}
\description{
This function generates a plot of reltionship between samples based on Hammering Distance
computed as the number of bins thats differ between samples
This function takes the path to the directory with CLONET output, as well as a tab separated file with
sample info and generates a plot of celularity and ploidy levels in the samples
}
