% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{process_maf}
\alias{process_maf}
\title{This function processes a MAF file
This function takes the path to a MAF file and processes the variant infomation
contained within it}
\usage{
process_maf(maf = "", sample_data = "", gen_data = "")
}
\arguments{
\item{maf}{Path to MAF file}

\item{sample_data}{Path to file with sample information}

\item{gen_data}{Path to file with gen to take into account}

\item{generate_file}{Generate an output file. Default TRUE}

\item{output_name}{Output file name without extension. Default "merged"}

\item{output_dir}{Path to output directory.}

\item{threads}{Number of threads to use.}
}
\description{
This function processes a MAF file
This function takes the path to a MAF file and processes the variant infomation
contained within it
}
