% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{process_variants}
\alias{process_variants}
\title{Process variants generated using variant_call and variant_call_strelka functions
This function takes the output directory of variant_call function and processes
the variants accordingly.}
\usage{
process_variants(
  bin_path = "tools/ensembl-vep/vep",
  bin_path2 = "tools/ensembl-vep/filter_vep",
  bin_path3 = "tools/bcftools/bcftools",
  bin_path4 = "tools/htslib/bgzip",
  bin_path5 = "tools/htslib/tabix",
  bin_path6 = "tools/vcf2maf/vcf2maf.pl",
  var_dir = "",
  bed_snps = "",
  ref_genome = "",
  output_dir = "",
  germ_pattern = "GL",
  vep_data = "~/.vep",
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to vep binary. Default tools/ensembl-vep/vep}

\item{bin_path2}{\link{REQUIRED} Path to vep filter binary. Default tools/ensembl-vep/filter_vep}

\item{bin_path3}{\link{REQUIRED} Path to bcftools binary. Default tools/bcftools/bcftools}

\item{bin_path4}{\link{REQUIRED} Path to bgzip. Default tools/htslib/bgzip}

\item{bin_path5}{\link{REQUIRED} Path to vaf2maf. Default tools/vaf2maf/vaf2maf.pl}

\item{var_dir}{\link{REQUIRED} Path to variant directory.}

\item{bed_snps}{\link{REQUIRED} Path to BED with Panel SNPs}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{germ_pattern}{\link{OPTIONAL} Pattern for normale samples}

\item{vep_data}{\link{REQUIRED} Path to VEP data directory.}

\item{verbose}{\link{DEFAULT==FALSE} Enables progress messages.}

\item{threads}{\link{OPTIONAL} Number of threads per job. Default 3}
}
\description{
Process variants generated using variant_call and variant_call_strelka functions
This function takes the output directory of variant_call function and processes
the variants accordingly.
}
