% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{split_vcf}
\alias{split_vcf}
\title{Splits multisample VCF file into multiple VCFs}
\usage{
split_vcf(
  bin_path = "tools/bcftools/bcftools",
  vcf = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{\link{Required} Path to bcftools binary. Default "tools/bcftools/bcftools"}

\item{vcf}{\link{Required} Path to VCF file to split.}

\item{verbose}{\link{Optional} Enables progress messages. Default False.}

\item{output_dir}{\link{Optional} Path to the output directory.}
}
\description{
This function takes a multisample VCF and splits into multiple VCFS for each sample
}
