% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tab_indx}
\alias{tab_indx}
\title{Index tab separated genomic regions}
\usage{
tab_indx(bin_path = "tools/htslib/tabix", file = "", verbose = FALSE)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to bcftools binary. Default tools/htslib/tabix.}

\item{file}{\link{REQUIRED} Path to VCF file to index.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
This function takes a .vcf.bgz tab separated genomic region file and generates an index for it
}
