% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vcf_bcftools}
\alias{vcf_bcftools}
\title{Variant calling using bcftools}
\usage{
vcf_bcftools(
  region = "",
  bin_path = "tools/bcftools/bcftools",
  bam = "",
  ref_genome = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{region}{Region to analyze. Optional}

\item{bin_path}{Path to fastQC executable. Default path tools/gatk/gatk.}

\item{bam}{Path to bam file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function calls variants in a sample.
}
