% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vcf_bcftools_parallel}
\alias{vcf_bcftools_parallel}
\title{Variant calling using bcftools on parallel per genomic region}
\usage{
vcf_bcftools_parallel(
  bin_path = "tools/bcftools/bcftools",
  bam = "",
  ref_genome = "",
  output_dir = "",
  region_bed = "",
  threads = 3,
  verbose = FALSE
)

vcf_bcftools_parallel(
  bin_path = "tools/bcftools/bcftools",
  bam = "",
  ref_genome = "",
  output_dir = "",
  region_bed = "",
  threads = 3,
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{Required} Path to bcftools binary. Default path tools/bcftools/bcftools.}

\item{bam}{\link{Required} Path to bam file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{output_dir}{Path to the output directory.}

\item{region_bed}{Path to bed file with regions to analyze.}

\item{threads}{\link{Optional} Number of threads. Default 3}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function calls SNPs in a sample

This function calls SNPs in a sample
}
