% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vcf_communality}
\alias{vcf_communality}
\title{VCF communality}
\usage{
vcf_communality(
  bin_path = "tools/bcftools/bcftools",
  vcf = "",
  vcf2 = "",
  output_name = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{vcf}{\link{REQUIRED} Path to VCF file}

\item{vcf2}{\link{REQUIRED} Path to VCF file}

\item{output_name}{\link{OPTIONAL} Name for output_files. IF not given it will use input vcf name}

\item{output_dir}{\link{OPTIONAL} Path to output directory. if not given outputs to current directory.}

\item{verbose}{\link{OPTIONAL} Extra verbose. Default FALSE}
}
\description{
Estimate \% of common SNPs between two VCF files and returns the communality between them
}
