% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vcf_concatenate}
\alias{vcf_concatenate}
\title{VCF file concatenation}
\usage{
vcf_concatenate(
  bin_path = "tools/bcftools/bcftools",
  vcf_dir = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{Path to bcftools binary. Default tools/bcftools/bcftools.}

\item{vcf_dir}{Path to directory with vcf files to concatenate.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function concatenates VCF files found in a directory.
}
