% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vcf_filter_variants}
\alias{vcf_filter_variants}
\title{VCF filtering using bcftools}
\usage{
vcf_filter_variants(
  bin_path = "tools/bcftools/bcftools",
  bin_path2 = "tools/htslib/bgzip",
  bin_path3 = "tools/htslib/tabix",
  unfil_vcf = "",
  qual = 30,
  mq = 40,
  state = "",
  ref = "",
  type = "",
  filter = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{Path to gatk binary. Default tools/gatk/gatk.}

\item{bin_path2}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path3}{Path to tabix binary. Default tools/htslib/tabix.}

\item{unfil_vcf}{Path to unfiltered vcf file.}

\item{qual}{Quality filter. Default 30.}

\item{mq}{Mapping quality filter. Default 40.}

\item{state}{\link{Optional} Variant state to select. Options: het/homo}

\item{ref}{\link{Optional} Filter variants with ID.}

\item{type}{\link{Optional} Variant type to include. Options: snp/indel. Default includes both}

\item{filter}{\link{Optional} Filter to include Options: PASS/.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function filters VCF calls using bcftools
}
