% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vcf_format}
\alias{vcf_format}
\title{VCF formating using bcftools}
\usage{
vcf_format(
  vcf = "",
  bin_path = "tools/bcftools/bcftools",
  bin_path2 = "tools/htslib/bgzip",
  bin_path3 = "tools/htslib/tabix",
 
    expr = "'\%CHROM\\\\t\%POS\\\\t\%ID\\\\t\%REF\\\\t\%ALT\\\\t\%QUAL\\\\t\%FILTER\\\\t\%INFO\\\\n'",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{vcf}{Path to vcf file.}

\item{bin_path}{Path to gatk binary. Default tools/gatk/gatk.}

\item{bin_path2}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path3}{Path to tabix binary. Default tools/htslib/tabix.}

\item{expr}{Expresion by which to format VCF. Default \%CHROM\\t\%POS\\t\%ID\\t\%REF\\t\%ALT\\t\%QUAL\\t\%FILTER\\t\%INFO\\n}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function formats VCF file using bcftools
}
