% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vcf_intersect_bed}
\alias{vcf_intersect_bed}
\title{Intersect vcf with bed file}
\usage{
vcf_intersect_bed(vcf = "", bed = "", output_name = "", output_dir = "")
}
\arguments{
\item{vcf}{\link{REQUIRED} Path to VCF file}

\item{bed}{\link{REQUIRED} Path to BED file}

\item{output_name}{\link{OPTIONAL} Name for output_files. IF not given it will use input vcf name}

\item{output_dir}{\link{OPTIONAL} Path to output directory. if not given outputs to current directory.}
}
\description{
Generate a set of variants covered by the regions found in/out the bed file
}
