% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vcf_platypus}
\alias{vcf_platypus}
\title{Variant calling using MuTECT2}
\usage{
vcf_platypus(
  bin_path = "tools/platypus/Platypus.py",
  tumor_bam = "",
  normal_bam = "",
  ref_genome = "",
  germ_resource = "",
  pon = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to fastQC executable. Default path tools/gatk/gatk.}

\item{tumor_bam}{Path to tumor bam file.}

\item{normal_bam}{Path to germline bam file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{germ_resource}{Path to germline resources vcf file.}

\item{pon}{\link{Optional} Path to panel of normal.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function calls somatic variants in a pair of tumor-normal matched samples, or
just in a tumor sample if no matched sample is not available.
}
