% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vcf_sets}
\alias{vcf_sets}
\title{Generate a set between a group of vcfs usign bcftools}
\usage{
vcf_sets(
  bin_path = "tools/bcftools/bcftools",
  vcf = "",
  vcf_dir = "",
  inset = TRUE,
  filter = "PASS",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to bcftools binary. Default tools/bcftools/bcftools}

\item{vcf}{\link{OPTIONAL} Path to a single vcf file or a vector of vcf files. Only if vcf_dir is not given.}

\item{vcf_dir}{\link{OPTIONAL} Path to a directory with vcf files to generate a set for. Only if vcf is not given.}

\item{inset}{\link{OPTIONAL} Generate and inset between. Default TRUE}

\item{filter}{\link{OPTIONAL} Filter variants by. Default PASS}

\item{output_dir}{\link{OPTIONAL} Path to output directory. Default current directory}

\item{verbose}{\link{Optional} Enables progress messages. Default False.}
}
\description{
This function takes a path to either a vcf file or to a directory of vcf files
and outputs an inset/outset between these two vcf files.
}
