% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vcf_sort}
\alias{vcf_sort}
\title{Sort VCF file}
\usage{
vcf_sort(
  bin_path = "tools/bcftools/bcftools",
  vcf = "",
  verbose = FALSE,
  output_dir = "",
  ram = 4,
  tmp_dir = ""
)
}
\arguments{
\item{bin_path}{\link{Required} Path to bcftools binary. Default tools/bcftools/bcftools.}

\item{vcf}{\link{Required} Path to VCF file to sort.}

\item{verbose}{\link{Optional} Enables progress messages. Default False.}

\item{output_dir}{\link{Optional} Path to the output directory. Default working directory.}

\item{ram}{\link{Required} RAM in GB to allocate sorting. Default 4Gb.}

\item{tmp_dir}{\link{Optional} Path to tmp file directory.}
}
\description{
This function takes a VCF file and sorts it genomic order
}
