% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vcf_stats_merge}
\alias{vcf_stats_merge}
\title{VCF file concatenation}
\usage{
vcf_stats_merge(
  bin_path = "tools/gatk/gatk",
  vcf_stats_dir = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{Path to gatk binary. Default tools/gatk/gatk.}

\item{vcf_stats_dir}{Path to directory with vcf stats files to merge.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function merges VCF stat files found in a directory.
}
