% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vep}
\alias{vep}
\title{Predict variant effect using VEP}
\usage{
vep(
  bin_path = "tools/ensembl-vep/vep",
  vcf = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{Path to bcftools binary. Default tools/bcftools/bcftools.}

\item{vcf}{Path to vcf file.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function predicts the effect of variant found in a VCF file
}
