% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_ASEQ}
\alias{call_ASEQ}
\title{Wrapper around ASEQ tool for pileup data.}
\usage{
call_ASEQ(
  bin_path = "tools/ASEQ/binaries/linux64/ASEQ",
  vcf = "",
  bam = "",
  mqr = "",
  mbq = "",
  mdc = "",
  htperc = "",
  pht = "",
  mode = "",
  output_dir = "",
  threads = 3,
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to ASEQ binary. Default tools/ASEQ/binaries/linux64/ASEQ}

\item{vcf}{\link{REQUIRED} Path to vcf file.}

\item{bam}{\link{REQUIRED} Path to bam file.}

\item{mqr}{\link{OPTIONAL} Filter by read mapping quality >=. Default value 1. See ASEQ doc.}

\item{mbq}{\link{OPTIONAL} Filter by base quality >=. Default value 1. See ASEQ doc.}

\item{mdc}{\link{OPTIONAL} Filter by coverage per base >=. Default value 1. See ASEQ doc.}

\item{htperc}{\link{OPTIONAL} Heterozigosity test based on percentage. Default value 0.2. Only in GENOTYPE mode. See ASEQ doc.}

\item{pht}{\link{OPTIONAL} P-value for heterozigosity test. Default value 0.2. Only in GENOTYPE mode. See ASEQ doc.}

\item{mode}{\link{OPTIONAL} Usage mode. Default value PILEUP. See ASEQ doc.7}

\item{output_dir}{\link{OPTIONAL}  Path to the output directory.}

\item{threads}{\link{OPTIONAL} Number of threads to use. Default value 3.}

\item{verbose}{\link{OPTIONAL}  Enables progress messages. Default False.}
}
\description{
This function takes a VCF and BAM files and generates a pileup for all the variants in the vcf file. See https://demichelislab.unitn.it/lib/exe/fetch.php?media=manual.pdf for ASEQ manual.
Two modes:
-PILEUP: Default use. Generates a pileup from the BAM file
-GENOTYPE: Generates genotype at VCF marked positions
The output is a ASEQ pileup format that can be used for downstream analysis. In addition, genotype mode generates a VCF file with all heterozygous calls.
}
