% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_mutect2_parallel}
\alias{call_mutect2_parallel}
\title{Variant calling using MuTECT2 on parallel per genomic region}
\usage{
call_mutect2_parallel(
  bin_path = "tools/gatk/gatk",
  bin_path2 = "tools/bcftools/bcftools",
  bin_path3 = "tools/htslib/bgzip",
  bin_path4 = "tools/htslib/tabix",
  tumor_bam = "",
  normal_bam = "",
  bam_dir = "",
  patient_id = "",
  germ_pattern = "GL",
  ref_genome = "",
  germ_resource = "",
  pon = "",
  output_dir = "",
  region_bed = "",
  db = "",
  interval = "",
  threads = 3,
  verbose = FALSE,
  chr_filter = "canonical",
  orientation = TRUE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk binary. Default path tools/gatk/gatk.}

\item{bin_path2}{\link{REQUIRED} Path to bcftools binary. Default path tools/bcftools/bcftools.}

\item{bin_path3}{\link{REQUIRED} Path to tabix binary. Default path tools/htslib/tabix.}

\item{bin_path4}{\link{REQUIRED} Path to bgzip binary. Default path tools/htslib/bgzip.}

\item{tumor_bam}{\link{OPTIONAL} Path to tumor bam file/s.}

\item{normal_bam}{\link{OPTIONAL} Path to germline bam file/s.}

\item{bam_dir}{\link{OPTIONAL} Path to bam_dir. Only if tumor_bam and/or normal_bam not given.}

\item{patient_id}{\link{OPTIONAL} Patients ID to identify project with. Otherwise name of tumor sample will be used.}

\item{germ_pattern}{\link{OPTIONAL} Identifier for normal samples. Default GL. Only used when bam_dir is given.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{germ_resource}{\link{REQUIRED} Path to germline resources vcf file.}

\item{pon}{\link{OPTIONAL} Path to panel of normal.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{region_bed}{\link{REQUIRED} Path to bed file with regions to analyze.}

\item{db}{\link{OPTIONAL} Path to vcf with known variants for contamination estimation.}

\item{interval}{\link{OPTIONAL} Path to vcf with intervals to analyze contamination.}

\item{threads}{\link{OPTIONAL} Number of threads. Default 3}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{chr_filter}{\link{OPTIONAL} Chromosomes to analyze. canonical/autosomal/all or a list of chromosomes}

\item{orientation}{\link{OPTIONAL} Generate a read orientation model to filter variants. Default False}
}
\description{
This function calls somatic variants in a pair of tumor-normal matched samples, or
just in a tumor sample if no matched sample is not available.
}
