% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_vep}
\alias{call_vep}
\title{Predict variant effect using VEP}
\usage{
call_vep(
  bin_path = "tools/ensembl-vep/vep",
  bin_path2 = "tools/htslib/bgzip",
  bin_path3 = "tools/htslib/tabix",
  vcf = "",
  verbose = FALSE,
  output_dir = "",
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to bcftools binary. Default tools/ensembl-vep/vep.}

\item{bin_path2}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path3}{Path to bgzip binary. Default tools/htslib/tabix.}

\item{vcf}{Path to vcf file.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}

\item{threads}{Number of threads to use. Default 3.}
}
\description{
This function predicts the effect of variant found in a VCF file
}
