% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{call_vep_maf}
\alias{call_vep_maf}
\title{Variant annotation in MAF format}
\usage{
call_vep_maf(
  bin_path = "tools/vcf2maf/vcf2maf.pl",
  vep_dir = "tools/ensembl/vep",
  vep_data = "~/.vep",
  vcf = "",
  verbose = FALSE,
  output_dir = "",
  patient_id = "",
  normal_id = "",
  ref_genome = ""
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to vcf2maf executable. Default path tools/vcf2maf/vcf2maf.pl}

\item{vep_dir}{\link{REQUIRED} Path to VEP executable directory. Default path tools/ensembl/vep}

\item{vep_data}{\link{REQUIRED} Path to VEP data directory. Default path ~/.vep}

\item{verbose}{\link{OPTIONAL}  Enables progress messages. Default False.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{vcf_overlay}{\link{REQUIRED} Path to vcf overlay to use as source.}

\item{threads}{\link{OPTIONAL}  Number of threads to use. Default 3.}

\item{targeted}{\link{OPTIONAL}  Sequencing data is exome/targeted. Default FALSE}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first sample in alphanumerical order will be used.}
}
\description{
This function annotates a VCF file using VEP, and then generates a
corresponding MAF file for it.
}
