% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{format_ASEQ_pileup}
\alias{format_ASEQ_pileup}
\title{VCF formating using bcftools}
\usage{
format_ASEQ_pileup(
  file = "",
  filter_indels = FALSE,
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{file}{Path to file to format.}

\item{filter_indels}{Filter indels from variants. Default false.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function formats ASEQ pileup file for downstream analysis.
}
