% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{format_ASEQ_pileup}
\alias{format_ASEQ_pileup}
\title{ASEQ pileup formatting for downstream analysis
This function formats ASEQ pileup file for downstream analysis.}
\usage{
format_ASEQ_pileup(file = "", verbose = FALSE, output_dir = "")
}
\arguments{
\item{file}{Path to file to format.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
ASEQ pileup formatting for downstream analysis
This function formats ASEQ pileup file for downstream analysis.
}
