% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{format_SNP_data}
\alias{format_SNP_data}
\title{This function filter and formats heterozygous SNP data for downstream analysis using clonet}
\usage{
format_SNP_data(
  bin_path = "tools/bcftools/bcftools",
  bin_path2 = "tools/htslib/bgzip",
  bin_path3 = "tools/htslib/tabix",
  bin_path4 = "tools/ASEQ/binaries/linux64/ASEQ",
  unfil_vcf_dir = "",
  bam_dir = "",
  qual = 30,
  mq = 40,
  min_cov = 20,
  patient_id = "",
  verbose = FALSE,
  output_dir = "",
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to gatk binary. Default tools/bcftools/bcftools.}

\item{bin_path2}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path3}{Path to tabix binary. Default tools/htslib/tabix.}

\item{bin_path4}{Path to ASEQ binary. Default tools/ASEQ/binaries/linux64/ASEQ}

\item{unfil_vcf_dir}{Path to unfiltered VCF file directory.}

\item{bam_dir}{Path to BAM file directory.}

\item{qual}{Quality filter. Default 30.}

\item{mq}{Mapping quality filter. Default 40.}

\item{min_cov}{Minimum coverage to filter. Default 20.}

\item{patient_id}{Patient identifier.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}

\item{threads}{Number of threads to use. Default 3}
}
\description{
This function takes a path of a directory of unfiltered VCFs files generated using Platypus and
the a path of the directory of BAMs from which those VCFs were generated.
}
