% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vcf_annotate}
\alias{vcf_annotate}
\title{VCF annotation using bcftools}
\usage{
vcf_annotate(
  bin_path = "tools/bcftools/bcftools",
  bin_path2 = "tools/htslib/bgzip",
  bin_path3 = "tools/htslib/tabix",
  vcf = "",
  db = "",
  verbose = FALSE,
  output_dir = "",
  threads = 1
)
}
\arguments{
\item{bin_path}{Path to gatk binary. Default tools/gatk/gatk.}

\item{bin_path2}{Path to bgzip binary. Default tools/htslib/bgzip.}

\item{bin_path3}{Path to tabix binary. Default tools/htslib/tabix.}

\item{vcf}{Path to vcf file.}

\item{db}{Path to vcf file with SNPs database.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}

\item{threads}{Additional in/out compression threads. Default 1.}

\item{state}{Variant state. Default het.}
}
\description{
This function annotates VCF file using bcftools
}
