% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vcf_filtering}
\alias{vcf_filtering}
\title{VCF filtering using GATK}
\usage{
vcf_filtering(
  bin_path = "tools/gatk/gatk",
  unfil_vcf = "",
  ref_genome = "",
  unfil_vcf_stats = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{Path to gatk binary. Default tools/gatk/gatk.}

\item{unfil_vcf}{Path to unfiltered vcf file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{unfil_vcf_stats}{Path to unfiltered vcf file stats.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function filters VCF calls using GATK statistics
}
