% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vcf_mutect2_parallel}
\alias{vcf_mutect2_parallel}
\title{Variant calling using MuTECT2 on parallel per genomic region}
\usage{
vcf_mutect2_parallel(
  region = "",
  bin_path = "tools/gatk/gatk",
  bin_path2 = "tools/bcftools/bcftools",
  tumor_bam = "",
  normal_bam = "",
  ref_genome = "",
  germ_resource = "",
  pon = "",
  output_dir = "",
  region_bed = "",
  threads = 3,
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{Required} Path to gatk binary. Default path tools/gatk/gatk.}

\item{bin_path2}{\link{Required} Path to bcftools binary. Default path tools/bcftools/bcftools.}

\item{tumor_bam}{\link{Required} Path to tumor bam file.}

\item{normal_bam}{Path to germline bam file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{germ_resource}{Path to germline resources vcf file.}

\item{pon}{\link{Optional} Path to panel of normal.}

\item{output_dir}{Path to the output directory.}

\item{region_bed}{Path to bed file with regions to analyze.}

\item{threads}{\link{Optional} Number of threads. Default 3}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function calls somatic variants in a pair of tumor-normal matched samples, or
just in a tumor sample if no matched sample is not available.
}
